+++ wled00/e131.cpp
@@ -119,7 +119,7 @@ void handleE131Packet(e131_packet_t* p, IPAddress clientIP, byte protocol){
   handleDMXData(uni, dmxChannels, e131_data, mde, previousUniverses);
 }
 
-void handleDMXData(uint16_t uni, uint16_t dmxChannels, uint8_t* e131_data, uint8_t mde, uint8_t previousUniverses) {
+void handleDMXData(uint16_t uni, uint16_t dmxChannels, uint8_t* e131_data, uint8_t mde, uint16_t previousUniverses) {
   #ifdef WLED_ENABLE_DMX
   // does not act on out-of-order packets yet
   if (e131ProxyUniverse > 0 && uni == e131ProxyUniverse) {
